Cleanup code Fix issue in the model persistence Remove periodic storage and store after analysing each request trace
