+++ components/global/core/io.cellery.observability.model.generator/src/main/java/io/cellery/observability/model/generator/internal/ServiceHolder.java
@@ -27,7 +27,6 @@ public class ServiceHolder {
     private static DataSourceService dataSourceService;
     private static ModelStoreManager modelStoreManager;
     private static ModelManager modelManager;
-    private static ModelPeriodicProcessor periodicProcessor;
 
     private ServiceHolder() {
     }
@@ -54,13 +53,5 @@ public class ServiceHolder {
 
     public static void setModelManager(ModelManager modelManager) {
         ServiceHolder.modelManager = modelManager;
-    }
-
-    public static ModelPeriodicProcessor getPeriodicProcessor() {
-        return periodicProcessor;
-    }
-
-    public static void setPeriodicProcessor(ModelPeriodicProcessor periodicProcessor) {
-        ServiceHolder.periodicProcessor = periodicProcessor;
     }
 }
